Refactor code and address a couple of sneaky bugs
